Fixes `FinderWindow` initialization
This commit fixes a bug that prevented `FinderWindow` from being
configured as the default class finder when the package was installed.

The bug was caused by using the old preference protocol, and assumed it
answered the #classFinder: message.

The bug was fixed by using the `PreferenceSet` protocol to configure
`FinderWindow` as the default class finder.